LUCENE-957: RAMDirectory fixed for more than Integer.MAX_VALUE bytes.
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@557678 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/store/RAMOutputStream.java
@@ -130,7 +130,7 @@ public class RAMOutputStream extends IndexOutput {
       currentBuffer = (byte[]) file.buffers.get(currentBufferIndex);
     }
     bufferPosition = 0;
-    bufferStart = BUFFER_SIZE * currentBufferIndex;
+    bufferStart = (long) BUFFER_SIZE * (long) currentBufferIndex;
     bufferLength = currentBuffer.length;
   }
